Add stack API and constructor implementations
This implements a stack as a doubly-linked list of "stack blocks" that
contain a larger chunk of memory. This is intended to avoid the
performance issues of having more sparsely allocated objects in the
stack by packing them closer together.

The constructors follow my usual pattern, alloc, init, and make. The
idea is that stack_alloc produces a correctly allocated pointer to a
stack object, that stack_init takes a pointer to an allocated stack
object and initializes it correctly, and that stack_make performs an
alloc then an init, for convenience.

The separation here lets you use a stack object allocated on the stack,
e.g:

    stack s;
    stack_init(&s);

You should only use stack_free on an object allocated with stack_alloc
or stack_make.